+++ components/header/IconHeader.tsx
@@ -4,7 +4,7 @@ import logo from "@/public/images/logos/sub-logo.svg";
 import Image from "next/image";
 import Link from "next/link";
 import React from "react";
-import Header from "./Header";
+import Header from "./components/Header";
 
 const IconHeader: React.FC<{ showUsername?: boolean }> = ({
   showUsername = true,
